fix(build): fix server actions in new Bun version
