+++ packages/brisa/src/utils/ast/constants.ts
@@ -3,3 +3,10 @@ export const FN = new Set([
   'FunctionExpression',
   'FunctionDeclaration',
 ]);
+
+// New versions of Bun (>= 1.1.33) if has a hash with the same name of the identifier
+// done in this Bun PR: https://github.com/oven-sh/bun/pull/14343
+export const jsxDEV = 'jsxDEV_7x81h0kn';
+export const jsx = 'jsx_w77yafs4';
+export const jsxs = 'jsxs_eh6c78nj';
+export const Fragment = 'Fragment_8vg9x3sq';
