--- packages/brisa/src/utils/is-jsx-indentifier/index.ts
@@ -1,5 +1,0 @@
-const JSX_REGEX = new RegExp(/^(jsx|jsxDEV|jsxs|_jsx|_jsxs)(_?[a-z0-9]+)?$/);
-
-export default function isJSXIdentifier(identifier: string) {
-  return JSX_REGEX.test(identifier);
-}
